Extract completion items into dedicated microcrate
