+++ crates/perl-lsp-completion/Cargo.toml
@@ -20,6 +20,7 @@ doctest = false
 perl-parser-core = { workspace = true }
 perl-semantic-analyzer = { workspace = true }
 perl-workspace-index = { workspace = true }
+perl-lsp-completion-items = { workspace = true }
 perl-keywords = { workspace = true }
 lsp-types = { version = "0.97.0", optional = true }
 url = "2.5.8"
